requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.1
pydantic_core==2.14.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
